requirements.txt
python-dotenv
pinecone-client
langchain-community
langchain-pinecone
langchain-openai
langchain
PyPDF2
tiktoken
requests
